Genetic algo introduction

+++ src/costfunctions/Nonlinearity.hh
@@ -6,7 +6,7 @@ namespace costfunctions {
 
 class Nonlinearity : public CostFunction {
 protected:
-    double evaluate(const std::vector<uint8_t>& input) const override;
+    double evaluate(const std::vector<int>& input) const override;
 };
 
 } // namespace costfunctions
